+++ jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/business/controller/admin/shippingInvoice/InvoiceController.java
@@ -84,6 +84,8 @@ public class InvoiceController {
 
     @Value("${jeecg.path.shippingInvoiceDetailDir}")
     private String INVOICE_DETAIL_DIR;
+    @Value("${jeecg.path.shippingInvoicePdfDir}")
+    private String INVOICE_PDF_DIR;
 
     @Autowired
     Environment env;
@@ -484,6 +486,7 @@ public class InvoiceController {
             if (metaData.getInvoiceCode().equals("error")) {
                 metaDataErrorList.add(metaData);
             } else {
+                filenameList.add(INVOICE_PDF_DIR + "//" + "Invoice N°" + metaData.getInvoiceCode() + " (" + metaData.getInvoiceEntity() + ").pdf");
                 filenameList.add(INVOICE_DIR + "//" + metaData.getFilename());
                 List<FactureDetail> factureDetails = shippingInvoiceService.getInvoiceDetail(metaData.getInvoiceCode());
                 List<SavRefundWithDetail> refunds = savRefundWithDetailService.getRefundsByInvoiceNumber(metaData.getInvoiceCode());
